requirements.txt
torch>=1.10.0
opencv-python
numpy
scikit-learn
insightface
onnxruntime-gpu
faiss-cpu
